% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline}
\alias{timeline}
\title{Timeline}
\usage{
timeline(title = NULL, subtitle = NULL, ...)
}
\arguments{
\item{title}{title of the timeline}

\item{subtitle}{subtitle of the timeline}

\item{...}{named lists containing the arguments and values to apply to the \code{mywidgets::timeline_item} function.}
}
\value{
A timeline
}
\description{
Create a timeline containing itens created with the \code{mywidgets::timeline_itens()} function.
}
\examples{
card(
  timeline(
    title = "Some title"
    ,subtitle = "Some subtitle"
    ,list(
      icon = "github"
      ,icon_color = "#C0B9DD"
      ,icon_size = "25px"
      ,value = 10
      ,value_color = "green"
      ,value_text = "commits in this week"
      ,secondary_value = "53\%"
      ,secondary_value_color = "red"
      ,secondary_value_text = "less than last month"
    )
    ,list(
      icon = "git"
      ,icon_color = "#030030"
      ,icon_size = "25px"
      ,value = 8
      ,value_color = "#404393"
      ,value_text = "repositories"
      ,secondary_value = 15
      ,secondary_value_color = "#123414"
      ,secondary_value_text = "forks"
    )
    ,list(
      icon = "lightbulb"
      ,icon_color = "#f14154"
      ,icon_size = "25px"
      ,value = 0
      ,value_color = "red"
      ,value_text = "good ideas in this year"
      ,secondary_value = "999\%"
      ,secondary_value_color = "blue"
      ,secondary_value_text = "more than yesterday"
    )
  )
)
}
