% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline_item}
\alias{timeline_item}
\title{Timeline item}
\usage{
timeline_item(
  icon = "github",
  icon_color = "#C0B9DD",
  icon_size = "25px",
  value = NULL,
  value_color = "#C0B9DD",
  value_text = NULL,
  text = NULL,
  secondary_value = NULL,
  secondary_value_color = "#C0B9DD",
  secondary_value_text = NULL
)
}
\arguments{
\item{icon}{a bootstrap icon string name}

\item{icon_color}{color of the icon}

\item{icon_size}{any valid css length unit}

\item{value}{main value of the timeline item}

\item{value_color}{color of the main value}

\item{value_text}{main value adjacent text}

\item{secondary_value}{secondary value of the timeline item}

\item{secondary_value_color}{color of the secondary value}

\item{secondary_value_text}{secondary value adjacent text}
}
\value{
A timeline item
}
\description{
Create an item containing a bootstrap icon, primary and secondary values and texts.
Each item will be appended to a timeline using the \code{mywidgets::timeline()} function.
}
