% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{valuebox}
\alias{valuebox}
\title{Valuebox}
\usage{
valuebox(
  icon = "github",
  icon_size = "1.6rem",
  icon_bg_color = "#C0B9DD",
  text = NULL,
  value = NULL,
  body_text = NULL,
  body_value = NULL,
  footer_text = NULL,
  footer_value = NULL,
  footer_value_color = "#A7F500"
)
}
\arguments{
\item{icon}{a bootstrap string name}

\item{icon_size}{any valid css length unit}

\item{icon_bg_color}{background color of the icon container}

\item{text}{header text}

\item{value}{main value}

\item{body_text}{card body text}

\item{body_value}{card body value}

\item{footer_text}{footer text}

\item{footer_value}{footer value}

\item{footer_value_color}{color of the value font}
}
\value{
A valuebox
}
\description{
Generate a \code{bslib::card()}-based valuebox
}
